services/users/requirements.txt
Flask==1.0.4
flask-sqlalchemy
psycopg2-binary
flask-testing
gunicorn
coverage
flask-cors
flask-migrate
flask-bcrypt
pyjwt